+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.4.0] - 2015-7-19