+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.1] - 2016-11-15
+### Fixed
+- Pull #16: Add support for non-string filter settings
+- Big thanks to [Stephen Young](https://github.com/hownowstephen) for the PR!
+
 ## [0.4.0] - 2015-7-19
 ### Added
 - IP Pools